+++ package.json
@@ -17,20 +17,20 @@
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.4",
     "@prisma/client": "^4.6.1",
-    "@trpc/client": "^9.27.2",
-    "@trpc/next": "^9.27.2",
-    "@trpc/react": "^9.27.2",
-    "@trpc/server": "^9.27.2",
+    "@tanstack/react-query": "^4.16.0",
+    "@trpc/client": "^10.0.0",
+    "@trpc/next": "^10.0.0",
+    "@trpc/react-query": "^10.0.0",
+    "@trpc/server": "^10.0.0",
     "chokidar": "^3.5.3",
     "classnames": "^2.3.1",
     "formidable": "^2.0.1",
     "jotai": "^1.8.1",
     "music-metadata": "^8.0.1",
-    "next": "12.3.0",
-    "next-auth": "^4.10.3",
+    "next": "13.0.2",
+    "next-auth": "^4.15.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-query": "^3.39.2",
     "sanitize-filename": "^1.6.3",
     "sharp": "^0.30.7",
     "superjson": "^1.9.1",
@@ -51,8 +51,8 @@
     "@types/ws": "^8.5.3",
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
-    "eslint": "8.22.0",
-    "eslint-config-next": "12.2.5",
+    "eslint": "^8.26.0",
+    "eslint-config-next": "13.0.2",
     "prisma": "^4.6.1",
     "typescript": "4.8.4"
   }
