+++ src/client/sw/useCachedTrack.ts
@@ -1,4 +1,4 @@
-import { useQuery } from "react-query"
+import { useQuery } from "@tanstack/react-query"
 
 export default function useCachedTrack({id, enabled}: {id?: string, enabled?: boolean}) {
 	const query = useQuery({
