+++ src/components/WatcherSocket/index.tsx
@@ -2,85 +2,86 @@ import { useEffect } from "react"
 import { env } from "env/client.mjs"
 import revalidateSwCache from "client/sw/revalidateSwCache"
 import { trpc } from "utils/trpc"
+import { useQueryClient } from "@tanstack/react-query"
 
 function onWsMessageData(data: any) {
 	if (data.type === "watcher:add") {
 		console.log("added track", data.payload)
-		revalidateSwCache("track.searchable")
-		revalidateSwCache("artist.searchable")
-		revalidateSwCache("album.searchable")
-		revalidateSwCache("genre.list")
+		revalidateSwCache(["track", "searchable"])
+		revalidateSwCache(["artist", "searchable"])
+		revalidateSwCache(["album", "searchable"])
+		revalidateSwCache(["genre", "list"])
 		if (data.payload?.artistId) {
-			revalidateSwCache("artist.miniature", {id: data.payload.artistId})
-			revalidateSwCache("artist.get", {id: data.payload.artistId})
-			revalidateSwCache("playlist.generate", { type: 'artist', id: data.payload.artistId })
+			revalidateSwCache(["artist", "miniature"], {id: data.payload.artistId})
+			revalidateSwCache(["artist", "get"], {id: data.payload.artistId})
+			revalidateSwCache(["playlist", "generate"], { type: 'artist', id: data.payload.artistId })
 		}
 		if (data.payload?.albumId) {
-			revalidateSwCache("album.miniature", {id: data.payload.albumId})
-			revalidateSwCache("album.get", {id: data.payload.albumId})
-			revalidateSwCache("playlist.generate", { type: 'album', id: data.payload.albumId })
-			revalidateSwCache("album.most-recent-add")
+			revalidateSwCache(["album", "miniature"], {id: data.payload.albumId})
+			revalidateSwCache(["album", "get"], {id: data.payload.albumId})
+			revalidateSwCache(["playlist", "generate"], { type: 'album', id: data.payload.albumId })
+			revalidateSwCache(["album", "mostRecentAdd"])
 		}
 	} else if (data.type === "watcher:add-playlist") {
 		console.log("added playlist")
-		revalidateSwCache("playlist.list")
-		revalidateSwCache("playlist.searchable")
+		revalidateSwCache(["playlist", "list"])
+		revalidateSwCache(["playlist", "searchable"])
 	} else if (data.type === "watcher:remove-playlist") {
 		console.log("removed playlist")
-		revalidateSwCache("playlist.list")
-		revalidateSwCache("playlist.searchable")
-		revalidateSwCache("playlist.get", { id: data.payload.id })
+		revalidateSwCache(["playlist", "list"])
+		revalidateSwCache(["playlist", "searchable"])
+		revalidateSwCache(["playlist", "get"], { id: data.payload.id })
 	} else if (data.type === "watcher:remove") {
 		console.log("removed", data.payload)
 		if (data.payload?.track) {
-			revalidateSwCache("track.searchable")
-			revalidateSwCache("track.miniature", {id: data.payload.track.id})
-			revalidateSwCache("playlist.generate", { type: 'track', id: data.payload.track.id })
+			revalidateSwCache(["track", "searchable"])
+			revalidateSwCache(["track", "miniature"], {id: data.payload.track.id})
+			revalidateSwCache(["playlist", "generate"], { type: 'track', id: data.payload.track.id })
 		}
 		if (data.payload?.artist) {
-			revalidateSwCache("artist.searchable")
-			revalidateSwCache("artist.miniature", {id: data.payload.artist.id})
-			revalidateSwCache("artist.get", {id: data.payload.artist.id})
-			revalidateSwCache("playlist.generate", { type: 'artist', id: data.payload.artist.id })
+			revalidateSwCache(["artist", "searchable"])
+			revalidateSwCache(["artist", "miniature"], {id: data.payload.artist.id})
+			revalidateSwCache(["artist", "get"], {id: data.payload.artist.id})
+			revalidateSwCache(["playlist", "generate"], { type: 'artist', id: data.payload.artist.id })
 		}
 		if (data.payload?.album) {
-			revalidateSwCache("album.searchable")
-			revalidateSwCache("album.miniature", {id: data.payload.album.id})
-			revalidateSwCache("album.get", {id: data.payload.album.id})
-			revalidateSwCache("playlist.generate", { type: 'album', id: data.payload.album.id })
+			revalidateSwCache(["album", "searchable"])
+			revalidateSwCache(["album", "miniature"], {id: data.payload.album.id})
+			revalidateSwCache(["album", "get"], {id: data.payload.album.id})
+			revalidateSwCache(["playlist", "generate"], { type: 'album', id: data.payload.album.id })
 		}
 		if (data.payload?.genre) {
-			revalidateSwCache("genre.list")
-			revalidateSwCache("genre.miniature", {id: data.payload.genre.id})
-			revalidateSwCache("genre.get", {id: data.payload.genre.id})
-			revalidateSwCache("playlist.generate", { type: 'genre', id: data.payload.genre.id })
+			revalidateSwCache(["genre", "list"])
+			revalidateSwCache(["genre", "miniature"], {id: data.payload.genre.id})
+			revalidateSwCache(["genre", "get"], {id: data.payload.genre.id})
+			revalidateSwCache(["playlist", "generate"], { type: 'genre', id: data.payload.genre.id })
 		}
 	} else if (data.type === "invalidate:track") {
 		console.log("invalidate track", data.payload)
-		revalidateSwCache("track.miniature", {id: data.payload.id})
+		revalidateSwCache(["track", "miniature"], {id: data.payload.id})
 	} else if (data.type === "invalidate:album") {
 		console.log("invalidate album", data.payload)
-		revalidateSwCache("album.miniature", {id: data.payload.id})
-		revalidateSwCache("album.get", {id: data.payload.id})
+		revalidateSwCache(["album", "miniature"], {id: data.payload.id})
+		revalidateSwCache(["album", "get"], {id: data.payload.id})
 	} else if (data.type === "invalidate:artist") {
 		console.log("invalidate artist", data.payload)
-		revalidateSwCache("artist.miniature", {id: data.payload.id})
-		revalidateSwCache("artist.get", {id: data.payload.id})
+		revalidateSwCache(["artist", "miniature"], {id: data.payload.id})
+		revalidateSwCache(["artist", "get"], {id: data.payload.id})
 	} else if (data.type === "invalidate:playlist") {
 		console.log("invalidate playlist", data.payload)
-		revalidateSwCache("playlist.get", {id: data.payload.id})
-		revalidateSwCache("playlist.list")
-		revalidateSwCache("playlist.searchable")
+		revalidateSwCache(["playlist", "get"], {id: data.payload.id})
+		revalidateSwCache(["playlist", "list"])
+		revalidateSwCache(["playlist", "searchable"])
 	} else if (data.type === "invalidate:listen-count") {
 		console.log("invalidate listen counts")
-		revalidateSwCache("artist.most-recent-listen")
-		revalidateSwCache("artist.least-recent-listen")
-		revalidateSwCache("album.most-recent-listen")
+		revalidateSwCache(["artist", "mostRecentListen"])
+		revalidateSwCache(["artist", "leastRecentListen"])
+		revalidateSwCache(["album", "mostRecentListen"])
 	} else if (data.type === "invalidate:likes") {
 		console.log("invalidate like counts")
-		revalidateSwCache("artist.most-fav")
-		revalidateSwCache("album.most-fav")
-		revalidateSwCache("genre.most-fav")
+		revalidateSwCache(["artist", "mostFav"])
+		revalidateSwCache(["album", "mostFav"])
+		revalidateSwCache(["genre", "mostFav"])
 	} else if (data.type === "global:message") {
 		// @ts-expect-error -- web socket messages aren't typed
 		console[data.payload?.level || 'log'](data.payload.message)
@@ -89,19 +90,20 @@ function onWsMessageData(data: any) {
 
 export default function WatcherSocket() {
 	const trpcClient = trpc.useContext()
+	const queryClient = useQueryClient()
 
 	useEffect(() => {
-		revalidateSwCache("track.searchable")
-		revalidateSwCache("artist.searchable")
-		revalidateSwCache("album.searchable")
-		revalidateSwCache("genre.list")
-		revalidateSwCache("artist.most-fav")
-		revalidateSwCache("artist.most-recent-listen")
-		revalidateSwCache("artist.least-recent-listen")
-		revalidateSwCache("album.most-fav")
-		revalidateSwCache("album.most-recent-listen")
-		revalidateSwCache("album.most-recent-add")
-		revalidateSwCache("genre.most-fav")
+		revalidateSwCache(["track", "searchable"])
+		revalidateSwCache(["artist", "searchable"])
+		revalidateSwCache(["album", "searchable"])
+		revalidateSwCache(["genre", "list"])
+		revalidateSwCache(["artist", "mostFav"])
+		revalidateSwCache(["artist", "mostRecentListen"])
+		revalidateSwCache(["artist", "leastRecentListen"])
+		revalidateSwCache(["album", "mostFav"])
+		revalidateSwCache(["album", "mostRecentListen"])
+		revalidateSwCache(["album", "mostRecentAdd"])
+		revalidateSwCache(["genre", "mostFav"])
 	}, [])
 
 	useEffect(() => {
@@ -156,20 +158,21 @@ export default function WatcherSocket() {
 				const message = event.data
 				if (message.type === 'sw-notify-when-track-cached') {
 					const id = message.payload.url.split('/').at(-1)
-					trpcClient.queryClient.invalidateQueries(['sw-cached-track', id])
+					queryClient.invalidateQueries(['sw-cached-track', id])
 				} else if (message.type === 'sw-trpc-invalidation') {
-					const queryKey = [message.payload.key]
-					if (message.payload.params) {
-						queryKey.push(message.payload.params)
+					const queryKey = message.payload.key
+					const queryParams = {
+						type: 'query',
+						...(message.payload.params ? {input: message.payload.params} : {}),
 					}
-					trpcClient.invalidateQueries(queryKey)
+					queryClient.invalidateQueries([queryKey, queryParams])
 				}
 			}, {signal: controller.signal})
 		})
 		return () => {
 			controller.abort()
 		}
-	}, [trpcClient])
+	}, [trpcClient, queryClient])
 
 	useEffect(() => {
 		const onOnline = async () => {
