--- src/server/router/context.ts
@@ -1,64 +1,0 @@
-// src/server/router/context.ts
-import * as trpc from "@trpc/server";
-import * as trpcNext from "@trpc/server/adapters/next";
-import {
-  Session,
-  unstable_getServerSession as getServerSession,
-} from "next-auth";
-import { authOptions as nextAuthOptions } from "pages/api/auth/[...nextauth]";
-import { prisma } from "server/db/client";
-import { fileWatcher } from "server/persistent/watcher";
-import { socketServer } from "server/persistent/ws";
-
-type CreateContextOptions = {
-  session: Session | null;
-};
-
-/** Use this helper for:
- * - testing, where we dont have to Mock Next.js' req/res
- * - trpc's `createSSGHelpers` where we don't have req/res
- **/
-export const createContextInner = async (opts: CreateContextOptions) => {
-  return {
-    session: opts.session,
-    prisma,
-    fileWatcher,
-    socketServer,
-  };
-};
-
-/**
- * This is the actual context you'll use in your router
- * @link https://trpc.io/docs/context
- **/
-export const createContext = async (
-  opts: trpcNext.CreateNextContextOptions,
-) => {
-  const session = await getServerSession(opts.req, opts.res, nextAuthOptions);
-
-  return await createContextInner({
-    session,
-  });
-};
-
-type Context = trpc.inferAsyncReturnType<typeof createContext>;
-
-export const createRouter = () => trpc.router<Context>();
-
-/**
- * Creates a tRPC router that asserts all queries and mutations are from an authorized user. Will throw an unauthorized error if a user is not signed in.
- */
- export function createProtectedRouter() {
-  return createRouter().middleware(({ ctx, next }) => {
-    if (!ctx.session || !ctx.session.user) {
-      throw new trpc.TRPCError({ code: "UNAUTHORIZED" });
-    }
-    return next({
-      ctx: {
-        ...ctx,
-        // infers that `session` is non-nullable to downstream resolvers
-        session: { ...ctx.session, user: ctx.session.user },
-      },
-    });
-  });
-}
