+++ src/types/next-auth.d.ts
@@ -1,4 +1,4 @@
-import { DefaultSession } from "next-auth";
+import { type DefaultSession } from "next-auth";
 
 declare module "next-auth" {
   /**
